Rename extension method
